requirements.txt
APScheduler==3.9.1
PyYAML==6.0
winotify==1.1.0
